+++ tests/default-extractor.test.js
@@ -476,6 +476,19 @@ test('classes in slim templates', async () => {
   expect(extractions).toContain('text-gray-500')
 })
 
+test("classes with fractional numeric values don't also generate the whole number utility", async () => {
+  const extractions = defaultExtractor(`
+    <div class="px-1.5 py-2.75">Hello world!</div>
+  `)
+
+  expect(extractions).toContain('px-1.5')
+  expect(extractions).toContain('py-2.75')
+  expect(extractions).not.toContain('px-1')
+  expect(extractions).not.toContain('5')
+  expect(extractions).not.toContain('py-2')
+  expect(extractions).not.toContain('75')
+})
+
 test('multi-word + arbitrary values + quotes', async () => {
   const extractions = defaultExtractor(`
     grid-cols-['repeat(2)']